refactor some components
