+++ src/components/MapLegendControl/index.tsx
@@ -1,0 +1,1 @@
+export { default as MapLegendControl } from "./MapLegendControl";
